+++ accounts/forms.py
@@ -1,0 +1,13 @@
+from django import forms
+from allauth.account.forms import SignupForm
+
+class CustomSignupForm(SignupForm):
+    first_name = forms.CharField(max_length=30, required=True, label="First Name")
+    last_name = forms.CharField(max_length=30, required=True, label="Last Name")
+
+    def save(self, request):
+        user = super(CustomSignupForm, self).save(request)
+        user.first_name = self.cleaned_data['first_name']
+        user.last_name = self.cleaned_data['last_name']
+        user.save()
+        return user