+++ genx/urls.py
@@ -4,5 +4,6 @@ from django.urls import path,include
 
 urlpatterns = [
     path('', include("core.urls")),
+    path('accounts/', include('allauth.urls')),
     path('admin/', admin.site.urls),
 ]